requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.1.7
django-allauth==0.44.0
django-appconf==1.0.4
django-bootstrap-pagination==1.7.1
django-bootstrap-v5==1.0.1
django-imagekit==4.0.2
idna==2.10
importlib-metadata==2.1.1
oauthlib==3.1.0
pilkit==2.0
Pillow==8.2.0
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
zipp==3.4.1
